python-service/requirements.txt
presidio-analyzer
presidio-anonymizer
spacy
transformers
torch
pymupdf
python-docx
pandas
numpy>=1.24.0
Pillow>=10.0.0
easyocr>=1.7.1
fastapi
uvicorn
python-multipart
python-dotenv
faker

# Optional: system Tesseract binary must also be installed (apt-get install tesseract-ocr)
# pytesseract
